bug fix - IndexOutOfRange

+++ src/main/java/com/alperenozcan/insurancenotebook/controller/CustomerController.java
@@ -1,17 +1,18 @@
 package com.alperenozcan.insurancenotebook.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.alperenozcan.insurancenotebook.entity.Customer;
+import com.alperenozcan.insurancenotebook.entity.CustomerHealthDetail;
+import com.alperenozcan.insurancenotebook.entity.InsuranceQuote;
 import com.alperenozcan.insurancenotebook.service.CustomerHealthDetailService;
 import com.alperenozcan.insurancenotebook.service.CustomerService;
 import com.alperenozcan.insurancenotebook.service.InsuranceQuoteService;
@@ -83,14 +84,25 @@ public class CustomerController {
 	@GetMapping("/delete")
 	public String deleteCustomer(@RequestParam("customerId") int theId) {
 		
-		// delete customer health details of the customer
-		int customerHealthDetailId = customerHealthDetailService.findByCustomerId(theId).getId();
-		customerHealthDetailService.deleteById(customerHealthDetailId);
+		// Delete customer's health detail if any 
+		Optional<CustomerHealthDetail> customerHealthDetail = customerHealthDetailService.findByCustomerId(theId);
+		if (customerHealthDetail.isPresent()) {
+			int customerHealthDetailId = customerHealthDetail.get().getId();
+			customerHealthDetailService.deleteById(customerHealthDetailId);	
+		}
 				
-		// delete insurance quotes of the customer if any
-		int insuranceQuotesId = insuranceQuoteService.findByCustomerId(theId).getId();
-		insuranceQuoteService.deleteById(insuranceQuotesId);
+				
+		// Delete customer's insurance quote(s) if any 
+		Optional<List<InsuranceQuote>> insuranceQuotes = insuranceQuoteService.findByCustomerId(theId);
+		if (insuranceQuotes.isPresent()) {
+			List<InsuranceQuote> list = insuranceQuotes.get();
+			for (InsuranceQuote quote : list){
+				insuranceQuoteService.deleteById(quote.getId());
+			}
+		}
 						
+		
+		
 		Customer tempCustomer = customerService.findById(theId);
 				
 		if (tempCustomer == null) {
